# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teranga GESCRIM API",
        "description": "API REST pour le système de Gestion de la Criminalité et de la Délinquance — Direction de la Sécurité Publique (DSP) du Sénégal.",
        "contact": {
            "name": "NAMASTECH",
            "email": "admin@gescrim.sn"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://gescrim.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/infractions": {
            "get": {
                "tags": [
                    "Infractions"
                ],
                "summary": "Liste des infractions",
                "operationId": "6e55860eab55d36fa726adcdd973f5c6",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commune_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "issue",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Constat\\u00e9e",
                                "D\\u00e9f\\u00e9r\\u00e9e"
                            ]
                        }
                    },
                    {
                        "name": "type_infraction_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sync_status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "synced"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Infractions"
                ],
                "summary": "Enregistrer une infraction",
                "description": "Crée une nouvelle infraction. Le champ sync_status peut être 'pending' pour les saisies offline.",
                "operationId": "afd0a32797f783bca664f687618d42ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type_infraction_id",
                                    "service_id",
                                    "annee",
                                    "date",
                                    "lieu",
                                    "commune_id",
                                    "issue"
                                ],
                                "properties": {
                                    "type_infraction_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-12"
                                    },
                                    "lieu": {
                                        "type": "string",
                                        "example": "Marché Sandaga, Dakar"
                                    },
                                    "commune_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "issue": {
                                        "type": "string",
                                        "example": "Constatée",
                                        "enum": [
                                            "Constatée",
                                            "Déférée"
                                        ]
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": 14.6928
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": -17.4467
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "sync_status": {
                                        "type": "string",
                                        "example": "synced",
                                        "enum": [
                                            "pending",
                                            "synced"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/infractions/{id}": {
            "get": {
                "tags": [
                    "Infractions"
                ],
                "summary": "Détail d'une infraction",
                "operationId": "10dd658f8b49fcb62a9e0337fef15ec5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Infractions"
                ],
                "summary": "Modifier une infraction",
                "description": "Règle métier : modification bloquée après 1 minute pour les agents.",
                "operationId": "731d3ade959ef5241b5344a368f383b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Infraction"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Délai dépassé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Infractions"
                ],
                "summary": "Supprimer une infraction",
                "operationId": "8531d91c6d982ebf32f09b16f412f8a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur rencontrée"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Infraction": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type_infraction_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "service_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "annee": {
                        "type": "integer",
                        "example": 2026
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-17"
                    },
                    "lieu": {
                        "type": "string",
                        "example": "Sandaga, Dakar"
                    },
                    "commune_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "issue": {
                        "type": "string",
                        "example": "Constatée"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 14.6928
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -17.4467
                    },
                    "description": {
                        "type": "string",
                        "example": "Usage de téléphone au volant constaté lors d'un contrôle."
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token JWT (sans le mot Bearer)"
            }
        }
    },
...
}